+++ targets/kit_nowcast_targets.R
@@ -27,7 +27,7 @@ kit_nowcast_targets <- list(
         predicted = value
       ) |>
       mutate(
-        model = "KIT simple nowcast original"
+        model = "original KIT simple nowcast"
       )
   ),
   tar_target(
@@ -68,7 +68,7 @@ kit_nowcast_targets <- list(
         predicted = value
       ) |>
       mutate(
-        model = "KIT simple nowcast revised"
+        model = "revised KIT simple nowcast"
       )
   ),
   tar_target(